+++ src/render/cdSimpleShader.h
@@ -68,8 +68,8 @@ namespace cdSimpleFullbrightPixel {
     unsigned int* GetPShader();  // @0x7D69A0
 }
 
-extern void nglDxRegisterVShader(unsigned int* VS, const unsigned int* Microcode);
-extern void nglDxRegisterPShader(unsigned int** PS, const unsigned int* Microcode);
+extern void nglDxRegisterVShader(unsigned long* VS, const unsigned int* Microcode);
+extern void nglDxRegisterPShader(unsigned long** PS, const unsigned int* Microcode);
 
 // ============================================================================
 // Externs
